Create EventSchedulerInterface class
Group event scheduler interface logic into subclass
